+++ funread_backend/TagsPerPage/views.py
@@ -10,12 +10,24 @@ from rest_framework.response import Response
 from rest_framework.decorators import api_view
 from rest_framework import status
 import hashlib
+import sys
+sys.path.append('funread_backend')
+import verifyJwt
+
 
 # Create your views here.
 
 #Metodo para mostrar todos los elementos de la lista TagsPerPage
 @ api_view(['GET'])
 def listed(request):
+
+    #token verification
+    authorization_header = request.headers.get('Authorization')
+    verify = verifyJwt.JWTValidator(authorization_header)
+    es_valido = verify.validar_token()
+    if es_valido==False:
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
+    
     tagsPerPage = TagsPerPage.objects.all()
     serializer = TagsPerPageSerializer (tagsPerPage, many=True)
     return Response(serializer.data)
@@ -23,6 +35,14 @@ def listed(request):
 #Metodo para buscar una variable por nombre
 @api_view(['GET'])
 def search(request):
+
+    #token verification
+    authorization_header = request.headers.get('Authorization')
+    verify = verifyJwt.JWTValidator(authorization_header)
+    es_valido = verify.validar_token()
+    if es_valido==False:
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
+    
     try:
         tagsPerPage  = TagsPerPage.objects.get(tagsPerPageId=request.data.get('tagsPerPageId'))
         print(tagsPerPage)
@@ -34,6 +54,14 @@ def search(request):
 #Metodo para agregar un elemento a la lista SharedBooks
 @api_view(['POST'])
 def add_new(request):
+
+    #token verification
+    authorization_header = request.headers.get('Authorization')
+    verify = verifyJwt.JWTValidator(authorization_header)
+    es_valido = verify.validar_token()
+    if es_valido==False:
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
+    
     print(request.data)
     data = {
         'pageId': request.data.get('pageId'),
@@ -48,6 +76,14 @@ def add_new(request):
 #Elimina un elemento de la lista SharedBooks
 @api_view(['DELETE'])
 def delete(request):
+
+    #token verification
+    authorization_header = request.headers.get('Authorization')
+    verify = verifyJwt.JWTValidator(authorization_header)
+    es_valido = verify.validar_token()
+    if es_valido==False:
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
+    
     tagsPerPage = TagsPerPage.objects.get(tagsPerPageId=request.data.get('tagsPerPageId'))
     tagsPerPage.delete()
     return Response({"msj":"Succesfully deleted"}, status=status.HTTP_200_OK)
@@ -56,6 +92,14 @@ def delete(request):
 #Metedo que cambia la variable de la lista SharedBooks
 @api_view(['PUT'])
 def update(request):
+
+    #token verification
+    authorization_header = request.headers.get('Authorization')
+    verify = verifyJwt.JWTValidator(authorization_header)
+    es_valido = verify.validar_token()
+    if es_valido==False:
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
+    
     tagsPerPage = TagsPerPage.objects.get(tagsPerPageId=request.data.get('tagsPerPageId'))
     serializer = TagsPerPageSerializer(tagsPerPage, data=request.data)
     if serializer.is_valid():
